+++ spec/lib/roadie/style_attribute_builder_spec.rb
@@ -4,11 +4,26 @@ module Roadie
   describe StyleAttributeBuilder do
     it "sorts the added properties" do
       builder = StyleAttributeBuilder.new
+
       builder << StyleProperty.new("color", "green", true, 1)
       builder << StyleProperty.new("font-size", "110%", false, 15)
       builder << StyleProperty.new("color", "red", false, 15)
 
-      expect(builder.attribute_string).to eq "color:red;font-size:110%;color:green !important"
+      expect(builder.attribute_string).to eq "font-size:110%;color:red;color:green !important"
+    end
+
+    it "preserves the order of added attributes with the same specificity" do
+      builder = StyleAttributeBuilder.new
+
+      builder << StyleProperty.new("color", "pink",  false, 50)
+      builder << StyleProperty.new("color", "red",   false, 50)
+      builder << StyleProperty.new("color", "green", false, 50)
+
+      # We need one different element to trigger the problem with Ruby's
+      # unstable sort
+      builder << StyleProperty.new("background", "white", false, 1)
+
+      expect(builder.attribute_string).to eq "background:white;color:pink;color:red;color:green"
     end
   end
 end